# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Let's Link API
  version: 1.0.0
  description: REST API for the Let's Link social planning platform.
servers:
  - url: http://localhost:4000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PartyMember:
      type: object
      properties:
        id:
          type: string
        party_id:
          type: string
        user_id:
          type: string
        status:
          type: string
          enum: [pending, confirmed, declined]
        location_lat:
          type: number
        location_lng:
          type: number
        location_name:
          type: string
        created_date:
          type: string
          format: date-time
        updated_date:
          type: string
          format: date-time
      required: [id, party_id, user_id]
...
    Photo:
      type: object
      properties:
        id:
          type: string
        party_id:
          type: string
        uploader_id:
          type: string
        file_url:
          type: string
          format: uri
        caption:
          type: string
        likes:
          type: array
          items:
            type: string
        created_date:
          type: string
          format: date-time
        updated_date:
          type: string
          format: date-time
      required: [id, party_id, uploader_id, file_url]
    Poll:
      type: object
      properties:
        id:
          type: string
        party_id:
          type: string
        created_by:
          type: string
        question:
          type: string
        options:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [open, closed]
        created_date:
          type: string
          format: date-time
        updated_date:
          type: string
          format: date-time
      required: [id, party_id, created_by, question, options]
...
paths:
...
  /party-members:
    get:
      security:
        - bearerAuth: []
      summary: List party members visible to the caller
      responses:
        '200':
          description: Members array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartyMember'
    post:
      security:
        - bearerAuth: []
      summary: Create a party member entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyMember'
      responses:
        '201':
          description: Party member created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyMember'
  /party-members/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      security:
        - bearerAuth: []
      summary: Update a party member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyMember'
    delete:
      security:
        - bearerAuth: []
      summary: Delete a party member
      responses:
        '204':
          description: Member removed
...
  /photos:
    get:
      security:
        - bearerAuth: []
      summary: List photos available to the caller
      responses:
        '200':
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photo'
    post:
      security:
        - bearerAuth: []
      summary: Upload metadata for a photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        '201':
          description: Photo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
  /photos/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      security:
        - bearerAuth: []
      summary: Update a photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
    delete:
      security:
        - bearerAuth: []
      summary: Delete a photo
      responses:
        '204':
          description: Photo deleted
  /polls:
    get:
      security:
        - bearerAuth: []
      summary: List polls for parties
      responses:
        '200':
          description: Poll list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Poll'
    post:
      security:
        - bearerAuth: []
      summary: Create a poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Poll'
      responses:
        '201':
          description: Poll created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
  /polls/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
    patch:
      security:
        - bearerAuth: []
      summary: Update a poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated poll
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Poll'
    delete:
      security:
        - bearerAuth: []
      summary: Delete a poll
      responses:
        '204':
          description: Poll deleted
...